Moved functions.
